ref(sandbox): Extract credentialed egress execution
Move the verified provider request path out of the sandbox egress proxy so the proxy owns only sandbox preflight concerns. This keeps the existing sandbox caller and behavior intact while giving the credentialed request path a smaller module boundary for future host/plugin egress work.

Refs GH-660
